Rename utils.numeric to utils.numerical
