+++ baybe/simulation.py
@@ -55,7 +55,7 @@ from baybe.searchspace import SearchSpaceType
 from baybe.targets.enum import TargetMode
 from baybe.utils.basic import set_random_seed
 from baybe.utils.dataframe import add_fake_results, add_parameter_noise
-from baybe.utils.numeric import closer_element, closest_element
+from baybe.utils.numerical import closer_element, closest_element
 
 if TYPE_CHECKING:
     from baybe.targets import NumericalTarget
