+++ baybe/utils/numerical.py
