+++ tests/hypothesis_strategies/parameters.py
@@ -16,7 +16,7 @@ from baybe.parameters.numerical import (
 )
 from baybe.parameters.substance import SubstanceEncoding, SubstanceParameter
 from baybe.utils.chemistry import get_canonical_smiles
-from baybe.utils.numeric import DTypeFloatNumpy
+from baybe.utils.numerical import DTypeFloatNumpy
 
 from .utils import interval
